replace meals card state with new one
