+++ app/src/commonMain/kotlin/com/maksimowiczm/foodyou/feature/settings/mealssettings/ui/MealsSettingsScreen.kt
@@ -1,9 +1,7 @@
 package com.maksimowiczm.foodyou.feature.settings.mealssettings.ui
 
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.imePadding
 import androidx.compose.foundation.layout.padding
@@ -11,9 +9,6 @@ import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ArrowBack
-import androidx.compose.material.icons.filled.Add
-import androidx.compose.material3.Card
-import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
@@ -29,17 +24,13 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.focus.FocusRequester
-import androidx.compose.ui.focus.focusRequester
 import androidx.compose.ui.input.nestedscroll.nestedScroll
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import com.maksimowiczm.foodyou.data.model.Meal
 import foodyou.app.generated.resources.*
 import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.launch
 import kotlinx.datetime.LocalTime
 import org.jetbrains.compose.resources.stringResource
 import org.koin.compose.viewmodel.koinViewModel
@@ -51,13 +42,16 @@ fun MealsSettingsScreen(
     viewModel: MealsSettingsViewModel = koinViewModel<MealsSettingsViewModel>()
 ) {
     val meals by viewModel.meals.collectAsStateWithLifecycle()
-
-    MealsSettingsScreen(
-        onBack = onBack,
+    val screenState = rememberMealsSettingsScreenState(
         meals = meals,
         onCreate = viewModel::createMeal,
         onUpdate = viewModel::updateMeal,
-        onDelete = viewModel::deleteMeal,
+        onDelete = viewModel::deleteMeal
+    )
+
+    MealsSettingsScreen(
+        onBack = onBack,
+        screenState = screenState,
         formatTime = viewModel::formatTime,
         modifier = modifier
     )
@@ -67,10 +61,7 @@ fun MealsSettingsScreen(
 @Composable
 private fun MealsSettingsScreen(
     onBack: () -> Unit,
-    meals: List<Meal>,
-    onCreate: suspend (name: String, from: LocalTime, to: LocalTime) -> Unit,
-    onUpdate: suspend (Meal) -> Unit,
-    onDelete: suspend (Meal) -> Unit,
+    screenState: MealsSettingsScreenState,
     formatTime: (LocalTime) -> String,
     modifier: Modifier = Modifier,
     coroutineScope: CoroutineScope = rememberCoroutineScope()
@@ -118,27 +109,16 @@ private fun MealsSettingsScreen(
             contentPadding = paddingValues
         ) {
             items(
-                items = meals,
+                items = screenState.meals,
                 key = { meal -> meal.id }
             ) { meal ->
-                val state = rememberMealsSettingsCardState(meal)
+                val state = screenState.rememberMealState(meal)
 
                 MealSettingsCard(
                     state = state,
                     showDeleteDialog = true,
-                    onDelete = {
-                        coroutineScope.launch {
-                            state.isLoading = true
-                            onDelete(meal)
-                        }
-                    },
-                    onConfirm = {
-                        coroutineScope.launch {
-                            state.isLoading = true
-                            onUpdate(state.intoMeal(meal.id))
-                            state.isLoading = false
-                        }
-                    },
+                    onDelete = state::onDelete,
+                    onConfirm = state::onUpdate,
                     formatTime = formatTime,
                     modifier = Modifier
                         .padding(horizontal = 16.dp)
@@ -151,17 +131,17 @@ private fun MealsSettingsScreen(
             item(
                 key = "create"
             ) {
-                CreateMealCard(
-                    creating = creating,
-                    onCreatingChange = { creating = it },
-                    onCreate = onCreate,
-                    formatTime = formatTime,
-                    modifier = Modifier
-                        .animateItem()
-                        .padding(horizontal = 16.dp)
-                        .focusRequester(createCardFocusRequester),
-                    coroutineScope = coroutineScope
-                )
+//                CreateMealCard(
+//                    creating = creating,
+//                    onCreatingChange = { creating = it },
+//                    onCreate = screenState::onCreate,
+//                    formatTime = formatTime,
+//                    modifier = Modifier
+//                        .animateItem()
+//                        .padding(horizontal = 16.dp)
+//                        .focusRequester(createCardFocusRequester),
+//                    coroutineScope = coroutineScope
+//                )
             }
 
             item {
@@ -171,53 +151,53 @@ private fun MealsSettingsScreen(
     }
 }
 
-@Composable
-private fun CreateMealCard(
-    creating: Boolean,
-    onCreatingChange: (Boolean) -> Unit,
-    onCreate: suspend (name: String, from: LocalTime, to: LocalTime) -> Unit,
-    formatTime: (LocalTime) -> String,
-    modifier: Modifier = Modifier,
-    coroutineScope: CoroutineScope = rememberCoroutineScope()
-) {
-    if (creating) {
-        val state = rememberMealsSettingsCardState()
-
-        MealSettingsCard(
-            state = state,
-            showDeleteDialog = false,
-            onDelete = { onCreatingChange(false) },
-            onConfirm = {
-                coroutineScope.launch {
-                    state.isLoading = true
-                    onCreate(state.nameInput.value, state.fromInput.value, state.toInput.value)
-                    state.isLoading = false
-                    onCreatingChange(false)
-                }
-            },
-            formatTime = formatTime,
-            modifier = modifier
-        )
-    } else {
-        Card(
-            onClick = { onCreatingChange(true) },
-            modifier = modifier,
-            colors = CardDefaults.outlinedCardColors(
-                containerColor = MealSettingsCardDefaults.colors().containerColor,
-                contentColor = MealSettingsCardDefaults.colors().contentColor
-            )
-        ) {
-            Box(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(16.dp),
-                contentAlignment = Alignment.Center
-            ) {
-                Icon(
-                    imageVector = Icons.Default.Add,
-                    contentDescription = stringResource(Res.string.action_add_meal)
-                )
-            }
-        }
-    }
-}
+// @Composable
+// private fun CreateMealCard(
+//    creating: Boolean,
+//    onCreatingChange: (Boolean) -> Unit,
+//    onCreate: suspend (name: String, from: LocalTime, to: LocalTime) -> Unit,
+//    formatTime: (LocalTime) -> String,
+//    modifier: Modifier = Modifier,
+//    coroutineScope: CoroutineScope = rememberCoroutineScope()
+// ) {
+//    if (creating) {
+//        val state = rememberMealsSettingsCardState()
+//
+//        MealSettingsCard(
+//            state = state,
+//            showDeleteDialog = false,
+//            onDelete = { onCreatingChange(false) },
+//            onConfirm = {
+//                coroutineScope.launch {
+//                    state.isLoading = true
+//                    onCreate(state.nameInput.value, state.fromInput.value, state.toInput.value)
+//                    state.isLoading = false
+//                    onCreatingChange(false)
+//                }
+//            },
+//            formatTime = formatTime,
+//            modifier = modifier
+//        )
+//    } else {
+//        Card(
+//            onClick = { onCreatingChange(true) },
+//            modifier = modifier,
+//            colors = CardDefaults.outlinedCardColors(
+//                containerColor = MealSettingsCardDefaults.colors().containerColor,
+//                contentColor = MealSettingsCardDefaults.colors().contentColor
+//            )
+//        ) {
+//            Box(
+//                modifier = Modifier
+//                    .fillMaxWidth()
+//                    .padding(16.dp),
+//                contentAlignment = Alignment.Center
+//            ) {
+//                Icon(
+//                    imageVector = Icons.Default.Add,
+//                    contentDescription = stringResource(Res.string.action_add_meal)
+//                )
+//            }
+//        }
+//    }
+// }
